+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Generate submission template.
 - Check that fixture-based tests actually ran.
 - Check types early, rather than waiting for file-not-found.
+- Generate JSON Schema from simpler Table Schema.